python-backend/requirements.txt
streamlit==1.51.0
langchain==1.2.0
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.2.3
langchain-google-genai==4.1.2
langchain-text-splitters==1.1.0
faiss-cpu==1.13.1
pypdf==6.4.0
pytesseract==0.3.13
python-dotenv==1.2.1
pillow==12.0.0
graphviz==0.21
google-generativeai==0.8.6
google-genai==1.56.0
fastapi
uvicorn
python-multipart
